FIX: starbus index (#325)
Fixes bug in PSSE parser during the creation of a starbus for 3-winding
transformers where the starbus number would keep increasing in
magnitude due to `find_max_bus_id` running at every iteration and
finding a new, higher max bus id. Fixes this by excluding any busname
with `starbus` in it, which is added to the name during the creation of
any starbus.